my-projects/Linkedin-Automation/requirements.txt
action-msgs==1.6.0
action-tutorials-interfaces==0.27.2
action-tutorials-py==0.27.2
actionlib-msgs==5.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
ajsonrpc==1.2.0
ament-cmake-test==2.0.7
ament-copyright==0.14.4
ament-cppcheck==0.14.4
ament-cpplint==0.14.4
ament-flake8==0.14.4
ament-index-python==1.5.2
ament-lint==0.14.4
ament-lint-cmake==0.14.4
ament-package==0.15.3
ament-pep257==0.14.4
ament-uncrustify==0.14.4
ament-xmllint==0.14.4
angles==1.16.0
anyio==4.6.2.post1
appdirs==1.4.4
argcomplete==1.8.1
async-timeout==5.0.1
attrs==25.1.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bitarray==2.9.3
bitstring==4.2.3
blinker==1.8.2
bloom==0.12.0
bottle==0.13.2
breezy==3.2.1
Brotli==1.0.9
builtin-interfaces==1.6.0
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloud-init==24.4
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.18.4
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.6
command-not-found==0.3
composition-interfaces==1.6.0
configobj==5.0.6
constantly==15.1.0
control-msgs==5.2.0
controller-manager==3.30.0
controller-manager-msgs==3.30.0
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cv-bridge==3.5.0
cycler==0.11.0
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
decorator==4.4.2
deep-translator==1.11.4
demo-nodes-py==0.27.2
Deprecated==1.2.13
diagnostic-msgs==5.0.1
diagnostic-updater==4.1.0
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.11.2
dulwich==0.20.31
ecdsa==0.19.0
empy==3.3.4
esptool==4.7.0
evdev==1.9.0
example-interfaces==0.10.2
examples-rclpy-executors==0.18.2
examples-rclpy-minimal-action-client==0.18.2
examples-rclpy-minimal-action-server==0.18.2
examples-rclpy-minimal-client==0.18.2
examples-rclpy-minimal-publisher==0.18.2
examples-rclpy-minimal-service==0.18.2
examples-rclpy-minimal-subscriber==0.18.2
exceptiongroup==1.2.2
fastbencode==0.0.5
fastimport==0.9.14
flake8==4.0.1
Flask==3.0.3
fonttools==4.29.1
frozenlist==1.5.0
fs==2.4.12
gast==0.5.2
geometry-msgs==5.0.1
googletrans==4.0.0rc1
gpg==1.16.0
gyp==0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
image-geometry==3.5.0
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
intelhex==2.3.0
interactive-markers==2.4.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.3.2
lark==1.1.1
laser-geometry==2.5.0
launch==2.0.4
launch-ros==0.24.2
launch-testing==2.0.4
launch-testing-ros==0.24.2
launch-xml==2.0.4
launch-yaml==2.0.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.6.0
logging-demo==0.27.2
lxml==4.8.0
lz4==3.1.3+dfsg
map-msgs==2.2.0
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.5.1
mccabe==0.6.1
mercurial==6.1.1
message-filters==4.7.1
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
multidict==6.1.0
nav-msgs==5.0.1
nest-asyncio==1.6.0
netifaces==0.11.0
nose2==0.9.2
notify2==0.3
numpy==2.1.1
oauthlib==3.2.0
olefile==0.46
osrf-pycommon==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
patiencediff==0.2.1
pcl-msgs==1.0.0
pendulum-msgs==0.27.2
pika==1.3.2
Pillow==9.0.1
platformio==6.1.16
plotly==5.24.1
pluggy==0.13.0
ply==3.11
propcache==0.2.1
psutil==5.9.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.8.0
pycurl==7.44.1
pydocstyle==6.1.1
pydot==1.4.2
pyelftools==0.31
pyflakes==2.4.0
pygame==2.6.1
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyNaCl==1.5.0
pynput==1.7.7
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gitlab==2.10.1
python-qt-binding==1.2.4
python-xlib==0.33
pythran==0.10.0
pytz==2024.2
PyYAML==5.4.1
qt-dotgraph==2.4.3
qt-gui==2.4.3
qt-gui-cpp==2.4.3
qt-gui-py-common==2.4.3
quality-of-service-demo-py==0.27.2
rcl-interfaces==1.6.0
rclpy==4.1.7
rcutils==6.2.4
redis==5.2.1
reedsolo==1.7.0
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==0.9.1
resource-retriever==3.2.3
retrying==1.3.4
rfc3986==1.5.0
rmw-dds-common==2.0.1
roman==3.3
ros-image-to-qimage==0.4.1
ros2action==0.25.8
ros2bag==0.22.8
ros2bag-mcap-cli==0.22.8
ros2bag-sqlite3-cli==0.22.8
ros2cli==0.25.8
ros2component==0.25.8
ros2doctor==0.25.8
ros2interface==0.25.8
ros2launch==0.24.2
ros2lifecycle==0.25.8
ros2multicast==0.25.8
ros2node==0.25.8
ros2param==0.25.8
ros2pkg==0.25.8
ros2run==0.25.8
ros2service==0.25.8
ros2topic==0.25.8
rosbag2-interfaces==0.22.8
rosbag2-py==0.22.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.6.0
rosidl-adapter==4.0.1
rosidl-cli==4.0.1
rosidl-cmake==4.0.1
rosidl-generator-c==4.0.1
rosidl-generator-cpp==4.0.1
rosidl-generator-py==0.18.0
rosidl-generator-type-description==4.0.1
rosidl-parser==4.0.1
rosidl-pycommon==4.0.1
rosidl-runtime-py==0.12.0
rosidl-typesupport-c==3.0.1
rosidl-typesupport-cpp==3.0.1
rosidl-typesupport-fastrtps-c==3.0.2
rosidl-typesupport-fastrtps-cpp==3.0.2
rosidl-typesupport-introspection-c==4.0.1
rosidl-typesupport-introspection-cpp==4.0.1
rospkg==1.5.1
rospkg-modules==1.5.1
rpyutils==0.3.2
rqt==1.3.4
rqt-action==2.1.2
rqt-bag==1.3.5
rqt-bag-plugins==1.3.5
rqt-console==2.1.1
rqt-controller-manager==3.30.0
rqt-gauges==0.0.3
rqt-graph==1.4.3
rqt-gui==1.3.4
rqt-gui-py==1.3.4
rqt-joint-trajectory-controller==3.28.0
rqt-moveit==1.0.1
rqt-msg==1.3.1
rqt-plot==1.2.3
rqt-publisher==1.6.3
rqt-py-common==1.3.4
rqt-py-console==1.1.1
rqt-reconfigure==1.3.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.6
rqt-robot-steering==1.0.0
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.1.1
rqt-shell==1.1.1
rqt-srv==1.1.1
rqt-tf-tree==1.0.5
rqt-topic==1.6.1
scipy==1.8.0
SecretStorage==3.3.1
selenium==4.28.1
semantic-version==2.10.0
sensor-msgs==5.0.1
sensor-msgs-py==5.0.1
service-identity==18.1.0
service-msgs==1.6.0
shape-msgs==5.0.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
spotipy==2.25.0
sros2==0.11.3
ssh-import-id==5.11
starlette==0.39.2
statistics-msgs==1.6.0
std-msgs==5.0.1
std-srvs==5.0.1
stereo-msgs==5.0.1
sympy==1.9
systemd-python==234
tabulate==0.9.0
teleop-twist-keyboard==2.4.0
tenacity==9.0.0
tf2-geometry-msgs==0.31.8
tf2-kdl==0.31.8
tf2-msgs==0.31.8
tf2-py==0.31.8
tf2-ros-py==0.31.8
tf2-sensor-msgs==0.31.8
tf2-tools==0.31.8
tk==0.1.0
toml==0.10.2
topic-monitor==0.27.2
trajectory-msgs==5.0.1
trio==0.29.0
trio-websocket==0.11.1
turtlesim==1.6.1
twilio==9.4.4
Twisted==22.1.0
type-description-interfaces==1.6.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unique-identifier-msgs==2.3.2
urllib3==2.2.3
uvicorn==0.30.6
vcstool==0.3.0
vcstools==0.1.42
visualization-msgs==5.0.1
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.13.3
wsproto==1.2.0
yarl==1.18.3
zipp==1.0.0
zope.interface==5.4.0
